Refactor HttpCollectorWireTests
